Break NextTableToSweepProvider API and fix Impl and BSI
