+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/sweep/SweeperTestSetup.java
@@ -112,7 +112,7 @@ public class SweeperTestSetup {
 
     protected void setNextTableToSweep(TableReference tableRef) {
         Mockito.doReturn(Optional.of(tableRef)).when(nextTableToSweepProvider)
-                .chooseNextTableToSweep(Mockito.any(), Mockito.anyLong());
+                .computeSweepPriorities(Mockito.any(), Mockito.anyLong());
     }
 
     protected void setupTaskRunner(SweepResults results) {
